Excise FileWatching by moving pidfile to Base
